requirements.txt
functions-framework
google-cloud-storage
google-cloud-pubsub
google-cloud-logging
flask